requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.3.1
python-docx==1.1.2
